Reuploaded Menu.h and Menu.cpp file

+++ SnakeGame/Food.cpp
@@ -3,7 +3,7 @@
 #include<time.h>
 
 
-sf::Vector2i Food::getCoordinates()
+sf::Vector2f Food::getCoordinates()
 {
 	return _coordinates;
 }
